fix build: speccing recent_projects that uses @ivars is tricky
Since the recent_projects method uses an instance variable, the spec has to
do its assert on the object the instance variable was modified. So when a project was
archived, it nullified the value of the @recent_projects of all users in the project.
The @user instance variable that designates the same database record but
is another ruby object has another @recent_projects variable that was not affected by the
nullification. Tricky.