Revert "Add a link to images on facebox to allow users download the original image."
Protip: Right click on an image in the browser; choose "Save Image As…". You're done.

If we're hyperlinking an image, people expect to see something new when they follow
the link—at least a bigger version of the image. Linking to the exact same image
that's currently displayed makes no sense.

This reverts commit bb6578e772a5648249d988726c682a40e8b114fb.