tasks: consolidate status/assigned user transitions
After the big comments & tasks refactoring, status changes are
independent of assigned user. This is now obvious in the UI too.
The only way that an assigned user can affect task status is that
it transitions from :new to :open automatically.

This change also keeps the initial status & assigned user in the
first comment if the task is created with one.