Issues with upload widget in pages caused by an irregularity that rails3 tightened up
Previously in Rails 2.x, setting the form tag's id/class when using
form_for was either overriden (or appended to) if passing in an object.

In rails3, if you specify a :id/:class key in html options, then that wins
and in the case of the upload widget this meant that the upload form no
longer had id='new_upload' on which the widget code depends to find the
form.