Always include all user's projects in _people js global
Fixes issue where assigned_id select box was occasionally empty.

The reason is that when you went to a specific project the _people
object was pòpulated for only that project. When switching to a
different project via ajax, _people wasn't updated and consequently
when Task.insertAssignableUsers() was called it couldn't find the users
for the project.