Moving rendering of GoogleDocs view to Thread view :(
The reason:

CommentForm's el is the actual form tag.
GoogleDocs.el contains other forms so it can't be rendered within the
comment form.
Changing CommentForm.el to be a holding div is too much work as a lot of
code is dependent on CommentForm.el being a form.
So forced to move rendering to Thread though logically it should be
rendered from CommentForm.