Fixing "reject searching in unauthorized project" api search spec
The api search controller uses a permission_to_search filter to filter
who can search. This method will return either 403 or 501 depending if
search is activated site-wide or not.

However, the belongs_to_project? filter kicks in before-hand and returns
a 403 if there's a current project and the user doesn't belong to it.